diveded SharePlace component into the PickLocation and PickImage components
