+++ src/screens/SharePlace/SharePlace.js
@@ -12,13 +12,11 @@ import { connect } from "react-redux";
 
 import { addPlace } from "../../store/actions/places";
 import DefaultInput from "../../components/UI/DefaultInput/DefaultInput";
+import PlaceInput from "../../components/PlaceInput/PlaceInput";
 import MainText from "../../components/UI/MainText/MainText";
 import HeadingText from "../../components/UI/HeadingText/HeadingText";
-
-const defaultImage = {
-  uri:
-    "http://ukrainetrek.com/blog/wp-content/uploads/2011/12/ukrainian-carpathians-landscape-16.jpg"
-};
+import PickImage from "../../components/PickImage/PickImage";
+import PickLocation from "../../components/PickLocation/PickLocation";
 
 class SharePlaceScreen extends Component {
   constructor(props) {
@@ -53,22 +51,12 @@ class SharePlaceScreen extends Component {
             <HeadingText>Share a Place with us!</HeadingText>
           </MainText>
 
-          <View style={styles.placeholder}>
-            <Image source={defaultImage} style={styles.image} />
-          </View>
+          <PickImage />
 
-          <View style={styles.button}>
-            <Button title="Pick Image" onPress={() => {}} />
-          </View>
+          <PickLocation />
 
-          <View style={styles.placeholder}>
-            <HeadingText>Map</HeadingText>
-          </View>
-          <View style={styles.button}>
-            <Button title="Locate Me" onPress={() => {}} />
-          </View>
+          <PlaceInput placeholder="Place Name" />
 
-          <DefaultInput placeholder="Place Name" />
           <View style={styles.button}>
             <Button title="Share the Place!" onPress={() => {}} />
           </View>
@@ -84,19 +72,8 @@ const styles = StyleSheet.create({
     flex: 1,
     alignItems: "center"
   },
-  placeholder: {
-    borderWidth: 1,
-    borderColor: "black",
-    backgroundColor: "#eee",
-    height: 150,
-    width: "80%"
-  },
   button: {
     margin: 10
-  },
-  image: {
-    width: "100%",
-    height: "100%"
   }
 });
